requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.4
python-dotenv==1.0.0
PyPDF2==3.0.1
